# docs/Architecture/features-docs-be/openapi-ADMIN-json.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hivek/admin/v1/kol-profiles": {
      "get": {
        "operationId": "KolProfileAdminController_findAll",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isVerified",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "categories",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Search/List KOL profiles",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      }
    },
    "/hivek/admin/v1/kol-profiles/{id}": {
      "get": {
        "operationId": "KolProfileAdminController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get KOL profile by ID",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      },
      "patch": {
        "operationId": "KolProfileAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKolProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update anything of an influencer (PATCH)",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      },
      "delete": {
        "operationId": "KolProfileAdminController_hardDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Hard delete KOL profile",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      }
    },
    "/hivek/admin/v1/kol-profiles/{id}/soft-delete": {
      "patch": {
        "operationId": "KolProfileAdminController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedBy",
            "required": false,
            "in": "query",
            "description": "The user performing the soft delete action",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "default": "system",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete KOL profile",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      }
    },
    "/hivek/admin/v1/kol-profiles/{id}/restore": {
      "patch": {
        "operationId": "KolProfileAdminController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Restore soft deleted KOL profile",
        "tags": [
          "ADMIN-kol-profiles"
        ]
      }
    },
...
    "/hivek/kol-profiles/{id}": {
      "patch": {
        "operationId": "TestKOLController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKolProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update anything of an influencer (PATCH)",
        "tags": [
          "test"
        ]
      }
    },
...
  },
  "info": {
    "title": "HiveK API",
    "description": "The API documentation for the HiveK Platform.\n\nNOTE: Sensitive endpoints (Sign-In, Sign-Up, OTP) are rate-limited to 5 requests per minute.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "UpdateKolProfileDto": {
        "type": "object",
        "properties": {
          "gender": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "scores": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {},
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": false,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "bio": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        }
      },
...
    }
  },
  "security": [
    {
      "x-api-key": []
    }
  ]
}